Added automatic syncer logic
